Remove hashCache type.
