+++ server/asset/eth/eth_test.go
@@ -21,6 +21,7 @@ import (
 	"decred.org/dcrdex/dex/encode"
 	dexeth "decred.org/dcrdex/dex/networks/eth"
 	swapv0 "decred.org/dcrdex/dex/networks/eth/contracts/v0"
+	"decred.org/dcrdex/server/asset"
 	"github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
@@ -200,7 +201,7 @@ func TestLoad(t *testing.T) {
 func TestDecodeCoinID(t *testing.T) {
 	drv := &Driver{}
 	txid := "0x1b86600b740d58ecc06eda8eba1c941c7ba3d285c78be89b56678da146ed53d1"
-	txHashB := mustDecodeHex("1b86600b740d58ecc06eda8eba1c941c7ba3d285c78be89b56678da146ed53d1")
+	txHashB := mustParseHex("1b86600b740d58ecc06eda8eba1c941c7ba3d285c78be89b56678da146ed53d1")
 
 	type test struct {
 		name    string
@@ -246,12 +247,6 @@ func TestRun(t *testing.T) {
 	backend := unconnectedETH(tLogger, new(config))
 	backend.node = &testNode{
 		bestHdr: &types.Header{Number: big.NewInt(1)},
-	}
-	backend.hashCache = &hashCache{
-		log:        tLogger,
-		signalMtx:  &backend.signalMtx,
-		blockChans: backend.blockChans,
-		node:       backend.node,
 	}
 	ch := backend.BlockChannel(1)
 	go func() {
@@ -724,10 +719,97 @@ func TestAccountBalance(t *testing.T) {
 	}
 }
 
-func mustDecodeHex(s string) []byte {
-	b, err := hex.DecodeString(s)
-	if err != nil {
-		panic("mustDecodeHex: " + err.Error())
+func TestPoll(t *testing.T) {
+	blkHdr := &types.Header{Number: big.NewInt(0)}
+	tests := []struct {
+		name                       string
+		bestHdr, hdrByHeight       *types.Header
+		bestHdrErr, hdrByHeightErr error
+		wantErr, preventSend       bool
+	}{{
+		name:    "ok nothing to do",
+		bestHdr: blkHdr,
+	}, {
+		name: "ok sequential",
+		bestHdr: &types.Header{
+			Number:     big.NewInt(1),
+			ParentHash: blkHdr.Hash(),
+		},
+	}, {
+		name: "ok fast blocks",
+		bestHdr: &types.Header{
+			Number: big.NewInt(1),
+		},
+		hdrByHeight: blkHdr,
+	}, {
+		name: "ok reorg",
+		bestHdr: &types.Header{
+			Number: big.NewInt(1),
+		},
+	}, {
+		name: "ok but cannot send",
+		bestHdr: &types.Header{
+			Number:     big.NewInt(1),
+			ParentHash: blkHdr.Hash(),
+		},
+		preventSend: true,
+	}, {
+		name:       "best header error",
+		bestHdrErr: errors.New(""),
+		wantErr:    true,
+	}, {
+		name: "header by height error",
+		bestHdr: &types.Header{
+			Number: big.NewInt(1),
+		},
+		hdrByHeightErr: errors.New(""),
+		wantErr:        true,
+	}}
+
+	for _, test := range tests {
+		node := &testNode{
+			bestHdr:        test.bestHdr,
+			bestHdrErr:     test.bestHdrErr,
+			hdrByHeight:    test.hdrByHeight,
+			hdrByHeightErr: test.hdrByHeightErr,
+		}
+		eth := &Backend{
+			log:        tLogger,
+			blockChans: make(map[chan *asset.BlockUpdate]struct{}),
+			node:       node,
+			bestHash: hashN{
+				hash: blkHdr.Hash(),
+			},
+		}
+		chSize := 1
+		if test.preventSend {
+			chSize = 0
+		}
+		ch := make(chan *asset.BlockUpdate, chSize)
+		eth.blockChans[ch] = struct{}{}
+		bu := new(asset.BlockUpdate)
+		wait := make(chan struct{})
+		go func() {
+			if test.preventSend {
+				close(wait)
+				return
+			}
+			select {
+			case bu = <-ch:
+			case <-time.After(time.Second * 2):
+			}
+			close(wait)
+		}()
+		eth.poll(nil)
+		<-wait
+		if test.wantErr {
+			if bu.Err == nil {
+				t.Fatalf("expected error for test %q", test.name)
+			}
+			continue
+		}
+		if bu.Err != nil {
+			t.Fatalf("unexpected error for test %q: %v", test.name, bu.Err)
+		}
 	}
-	return b
 }
